Move admin keyboards to their own packages
