+++ handlers/admin/backups/backups_handler.py
@@ -3,7 +3,7 @@ from aiogram.types import CallbackQuery
 
 from backup import backup_database
 from filters.admin import IsAdminFilter
-from keyboards.admin.panel_kb import AdminPanelCallback, build_admin_back_kb
+from ..panel.keyboard import AdminPanelCallback, build_admin_back_kb
 
 
 router = Router()
