+++ handlers/admin/panel/keyboard.py
