+++ handlers/admin/panel/panel_handler.py
@@ -5,7 +5,7 @@ from aiogram.types import CallbackQuery, Message
 
 from bot import version
 from filters.admin import IsAdminFilter
-from keyboards.admin.panel_kb import AdminPanelCallback, build_management_kb, build_panel_kb
+from keyboard import AdminPanelCallback, build_management_kb, build_panel_kb
 
 
 router = Router()
