+++ handlers/admin/sender/keyboard.py
@@ -2,7 +2,7 @@ from aiogram.filters.callback_data import CallbackData
 from aiogram.types import InlineKeyboardMarkup
 from aiogram.utils.keyboard import InlineKeyboardBuilder
 
-from keyboards.admin.panel_kb import build_admin_back_btn
+from ..panel.keyboard import build_admin_back_btn
 
 
 class AdminSenderCallback(CallbackData, prefix="admin_sender"):
