+++ handlers/admin/users/keyboard.py
@@ -5,7 +5,7 @@ from aiogram.types import InlineKeyboardButton, InlineKeyboardMarkup
 from aiogram.utils.keyboard import InlineKeyboardBuilder
 from config import RENEWAL_PRICES
 
-from keyboards.admin.panel_kb import build_admin_back_btn
+from ..panel.keyboard import build_admin_back_btn
 
 
 class AdminUserEditorCallback(CallbackData, prefix="admin_users"):
